// apps/api/src/generated/kysely/valuations/LegalEntityType.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Represents the enum valuations.LegalEntityType */
enum LegalEntityType {
  COMPANY = 'COMPANY',
  ESOP = 'ESOP',
  FUND = 'FUND',
  NATURAL_PERSON = 'NATURAL_PERSON',
  PORTFOLIO_COMPANY = 'PORTFOLIO_COMPANY',
  SPV = 'SPV',
};

export default LegalEntityType;